\name{nirs4all_domains}
\alias{nirs4all_domains}
\alias{formats}
\alias{io}
\alias{datasets}
\alias{methods}
\alias{dag_ml}
\alias{dag_ml_data}
\title{Access upstream domain namespaces}
\usage{
formats()

io()

datasets()

methods()

dag_ml()

dag_ml_data()
}
\value{The loaded upstream R package namespace.}
\description{Compatibility accessors moved from the former nirs4all-core R facade.}
